Revert "Add in-game discord user colors, message updates, and new line indicators (#29)"
This reverts commit 31f2b7e995f92f7c95c3f3bcbded9c3f437c5865.
